Use html tag for metadata message
- remove old footprint version based on chartjs pie
