+++ web/frontend/src/Job.root.svelte
@@ -27,7 +27,7 @@
     import TagManagement from "./TagManagement.svelte";
     import MetricSelection from "./MetricSelection.svelte";
     import StatsTable from "./StatsTable.svelte";
-    import JobFootprintBars from "./JobFootprintBars.svelte";
+    import JobFootprint from "./JobFootprint.svelte";
     import { getContext } from "svelte";
 
     export let dbid;
@@ -206,12 +206,7 @@
     {#if $jobMetrics.data}
         {#key $jobMetrics.data}
             <Col>
-                <!-- <JobFootprint
-                    bind:this={jobFootprint}
-                    job={$initq.data.job}
-                    jobMetrics={$jobMetrics.data.jobMetrics}
-                /> -->
-                <JobFootprintBars
+                <JobFootprint
                     bind:this={jobFootprint}
                     job={$initq.data.job}
                     jobMetrics={$jobMetrics.data.jobMetrics}
